fix: bump litertlm 0.9.0-alpha01 → 0.9.0 stable; cast Float → Double
0.9.0-alpha01 tag doesn't exist on GitHub/Maven; the real API
is defined in the stable 0.9.0 release (sealed class Backend with
Backend.GPU()/CPU() and Message.user() factory).

Also add .toDouble() casts on topP and temperature since
SamplerConfig takes Double, not Float.

+++ app/src/main/java/com/chatgemma/app/ai/LiteRtInferenceEngine.kt
@@ -55,8 +55,8 @@ class LiteRtInferenceEngine @Inject constructor(
             try {
                 val samplerConfig = SamplerConfig(
                     topK = params.topK,
-                    topP = params.topP,
-                    temperature = params.temperature
+                    topP = params.topP.toDouble(),
+                    temperature = params.temperature.toDouble()
                 )
                 val convConfig = ConversationConfig(samplerConfig = samplerConfig)
                 val conversation = engine!!.createConversation(convConfig)
